# pkg/api/api/openapi/apiserver/v1/apiserver.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "eino-show API",
    "version": "1.0",
    "contact": {
      "name": "eino-show 项目",
      "url": "https://github.com/ashwinyue/eino-show",
      "email": "stary99c@163.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/ashwinyue/eino-show/blob/master/LICENSE"
    }
  },
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/agents/builtin": {
      "get": {
        "summary": "获取内置 Agent 列表",
        "operationId": "ListBuiltinAgents",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListBuiltinAgentsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Agent 管理"
        ]
      }
    },
    "/v1/custom-agents": {
      "get": {
        "summary": "获取 Agent 列表",
        "operationId": "ListAgents",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAgentsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Agent 管理"
        ]
      },
      "post": {
        "summary": "创建 Agent",
        "operationId": "CreateAgent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateAgentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateAgentRequest"
            }
          }
        ],
        "tags": [
          "Agent 管理"
        ]
      }
    },
    "/v1/custom-agents/{id}": {
      "get": {
        "summary": "获取 Agent 详情",
        "operationId": "GetAgent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAgentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Agent 管理"
        ]
      },
      "delete": {
        "summary": "删除 Agent",
        "operationId": "DeleteAgent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteAgentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Agent 管理"
        ]
      },
      "put": {
        "summary": "更新 Agent",
        "operationId": "UpdateAgent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateAgentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MiniBlogUpdateAgentBody"
            }
          }
        ],
        "tags": [
          "Agent 管理"
        ]
      }
    },
    "/v1/knowledge-bases": {
      "get": {
        "summary": "获取知识库列表",
        "operationId": "ListKnowledgeBases",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListKnowledgeBasesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "知识库管理"
        ]
      },
      "post": {
        "summary": "创建知识库",
        "operationId": "CreateKnowledgeBase",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateKnowledgeBaseResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateKnowledgeBaseRequest"
            }
          }
        ],
        "tags": [
          "知识库管理"
        ]
      }
    },
    "/v1/knowledge-bases/{id}": {
      "get": {
        "summary": "获取知识库详情",
        "operationId": "GetKnowledgeBase",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetKnowledgeBaseResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "知识库管理"
        ]
      },
      "delete": {
        "summary": "删除知识库",
        "operationId": "DeleteKnowledgeBase",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteKnowledgeBaseResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "知识库管理"
        ]
      },
      "put": {
        "summary": "更新知识库",
        "operationId": "UpdateKnowledgeBase",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateKnowledgeBaseResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MiniBlogUpdateKnowledgeBaseBody"
            }
          }
        ],
        "tags": [
          "知识库管理"
        ]
      }
    },
    "/v1/knowledge-bases/{id}/stats": {
      "get": {
        "summary": "获取知识库统计",
        "operationId": "GetKnowledgeStats",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1KnowledgeStatsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "知识库管理"
        ]
      }
    },
    "/v1/knowledge-bases/{kbId}/knowledge": {
      "get": {
        "summary": "获取知识列表",
        "operationId": "ListKnowledges",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListKnowledgesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "知识库管理"
        ]
      }
    },
    "/v1/knowledge/{id}": {
      "delete": {
        "summary": "删除知识",
        "operationId": "DeleteKnowledge",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteKnowledgeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "知识库管理"
        ]
      }
    },
...
    "/v1/sessions": {
      "get": {
        "summary": "获取会话列表",
        "operationId": "ListSessions",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListSessionsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "会话管理"
        ]
      },
      "post": {
        "summary": "创建会话",
        "operationId": "CreateSession",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateSessionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateSessionRequest"
            }
          }
        ],
        "tags": [
          "会话管理"
        ]
      }
    },
    "/v1/sessions/{id}": {
      "get": {
        "summary": "获取会话详情",
        "operationId": "GetSession",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetSessionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "会话管理"
        ]
      },
      "delete": {
        "summary": "删除会话",
        "operationId": "DeleteSession",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteSessionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "会话管理"
        ]
      },
      "put": {
        "summary": "更新会话",
        "operationId": "UpdateSession",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateSessionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MiniBlogUpdateSessionBody"
            }
          }
        ],
        "tags": [
          "会话管理"
        ]
      }
    },
...
  },
  "definitions": {
...
    "MiniBlogUpdateAgentBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "config": {
          "$ref": "#/definitions/v1CustomAgentConfig"
        }
      },
      "title": "UpdateAgentRequest 更新 Agent 请求"
    },
    "MiniBlogUpdateKnowledgeBaseBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "chunkingConfig": {
          "$ref": "#/definitions/v1KnowledgeChunkingConfig"
        },
        "imageProcessingConfig": {
          "$ref": "#/definitions/v1KnowledgeImageConfig"
        },
        "embeddingModelId": {
          "type": "string"
        },
        "summaryModelId": {
          "type": "string"
        },
        "rerankModelId": {
          "type": "string"
        }
      },
      "title": "UpdateKnowledgeBaseRequest 更新知识库请求"
    },
...
    "MiniBlogUpdateSessionBody": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "agentConfig": {
          "$ref": "#/definitions/v1SessionAgentConfig"
        },
        "contextConfig": {
          "$ref": "#/definitions/v1SessionContextConfig"
        }
      },
      "title": "UpdateSessionRequest 更新会话请求"
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Agent": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "isBuiltin": {
          "type": "boolean"
        },
        "config": {
          "$ref": "#/definitions/v1CustomAgentConfig"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "Agent 自定义 Agent"
    },
    "v1BuiltinAgent": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "title": "BuiltinAgent 内置 Agent"
    },
...
    "v1CreateAgentRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "config": {
          "$ref": "#/definitions/v1CustomAgentConfig"
        }
      },
      "title": "CreateAgentRequest 创建 Agent 请求"
    },
    "v1CreateAgentResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "CreateAgentResponse 创建 Agent 响应"
    },
    "v1CreateKnowledgeBaseRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "chunkingConfig": {
          "$ref": "#/definitions/v1KnowledgeChunkingConfig"
        },
        "imageProcessingConfig": {
          "$ref": "#/definitions/v1KnowledgeImageConfig"
        },
        "embeddingModelId": {
          "type": "string"
        },
        "summaryModelId": {
          "type": "string"
        },
        "rerankModelId": {
          "type": "string"
        }
      },
      "title": "CreateKnowledgeBaseRequest 创建知识库请求"
    },
    "v1CreateKnowledgeBaseResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "CreateKnowledgeBaseResponse 创建知识库响应"
    },
...
    "v1CreateSessionRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "agentId": {
          "type": "string"
        },
        "agentConfig": {
          "$ref": "#/definitions/v1SessionAgentConfig"
        },
        "contextConfig": {
          "$ref": "#/definitions/v1SessionContextConfig"
        }
      },
      "title": "CreateSessionRequest 创建会话请求"
    },
    "v1CreateSessionResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "CreateSessionResponse 创建会话响应"
    },
...
    "v1CustomAgentConfig": {
      "type": "object",
      "properties": {
        "instruction": {
          "type": "string"
        },
        "temperature": {
          "type": "number",
          "format": "double"
        },
        "maxIterations": {
          "type": "integer",
          "format": "int32"
        },
        "tools": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "knowledgeBases": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "modelId": {
          "type": "string"
        },
        "enableWebSearch": {
          "type": "boolean"
        }
      },
      "title": "CustomAgentConfig Agent 配置"
    },
    "v1DeleteAgentResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "DeleteAgentResponse 删除 Agent 响应"
    },
    "v1DeleteKnowledgeBaseResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "DeleteKnowledgeBaseResponse 删除知识库响应"
    },
    "v1DeleteKnowledgeResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "DeleteKnowledgeResponse 删除知识项响应"
    },
...
    "v1DeleteSessionResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "DeleteSessionResponse 删除会话响应"
    },
...
    "v1GetAgentResponse": {
      "type": "object",
      "properties": {
        "agent": {
          "$ref": "#/definitions/v1Agent"
        }
      },
      "title": "GetAgentResponse 获取 Agent 响应"
    },
    "v1GetKnowledgeBaseResponse": {
      "type": "object",
      "properties": {
        "knowledgeBase": {
          "$ref": "#/definitions/v1KnowledgeBase"
        }
      },
      "title": "GetKnowledgeBaseResponse 获取知识库响应"
    },
...
    "v1GetSessionResponse": {
      "type": "object",
      "properties": {
        "session": {
          "$ref": "#/definitions/v1Session"
        }
      },
      "title": "GetSessionResponse 获取会话响应"
    },
...
    "v1KnowledgeBase": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "chunkingConfig": {
          "$ref": "#/definitions/v1KnowledgeChunkingConfig"
        },
        "imageProcessingConfig": {
          "$ref": "#/definitions/v1KnowledgeImageConfig"
        },
        "embeddingModelId": {
          "type": "string"
        },
        "summaryModelId": {
          "type": "string"
        },
        "rerankModelId": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "KnowledgeBase 知识库"
    },
    "v1KnowledgeChunkingConfig": {
      "type": "object",
      "properties": {
        "chunkSize": {
          "type": "integer",
          "format": "int32"
        },
        "chunkOverlap": {
          "type": "integer",
          "format": "int32"
        },
        "splitMarkers": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "keepSeparator": {
          "type": "boolean"
        }
      },
      "title": "KnowledgeChunkingConfig 文档分块配置"
    },
...
    "v1KnowledgeStatsResponse": {
      "type": "object",
      "properties": {
        "knowledgeCount": {
          "type": "string",
          "format": "int64"
        },
        "chunkCount": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "KnowledgeStatsResponse 知识库统计响应"
    },
    "v1ListAgentsResponse": {
      "type": "object",
      "properties": {
        "agents": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Agent"
          }
        },
        "total": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "ListAgentsResponse 获取 Agent 列表响应"
    },
    "v1ListBuiltinAgentsResponse": {
      "type": "object",
      "properties": {
        "agents": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1BuiltinAgent"
          }
        }
      },
      "title": "ListBuiltinAgentsResponse 获取内置 Agent 列表响应"
    },
    "v1ListKnowledgeBasesResponse": {
      "type": "object",
      "properties": {
        "knowledgeBases": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1KnowledgeBase"
          }
        },
        "total": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "ListKnowledgeBasesResponse 获取知识库列表响应"
    },
    "v1ListKnowledgesResponse": {
      "type": "object",
      "properties": {
        "knowledge": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Knowledge"
          }
        },
        "total": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "ListKnowledgesResponse 获取知识列表响应"
    },
...
    "v1ListSessionsResponse": {
      "type": "object",
      "properties": {
        "sessions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Session"
          }
        },
        "total": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "ListSessionsResponse 获取会话列表响应"
    },
...
    "v1UpdateAgentResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "UpdateAgentResponse 更新 Agent 响应"
    },
    "v1UpdateKnowledgeBaseResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "UpdateKnowledgeBaseResponse 更新知识库响应"
    },
...
    "v1UpdateSessionResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "UpdateSessionResponse 更新会话响应"
    },
...
  }
}